Move THD class to main_THD.py
