+++ Asterix/optics/__init__.py
@@ -5,7 +5,6 @@ from .pupil import *
 from .coronagraph import *
 from .deformable_mirror import *
 from .testbed import *
-from .thd import *
 
 from .phase_amplitude_functions import *
 from .propagation_functions import *
